Fix NFC Forum T5T Tag 130KB Read/Write failure.
Bug: 309749099
Test: T5T RW test cases with 130KB NDEF data

+++ src/nfc/include/rw_api.h
@@ -1279,7 +1279,7 @@ extern tNFC_STATUS RW_I93ReadNDef(void);
 **                  NFC_STATUS_FAILED if I93 is busy or other error
 **
 *******************************************************************************/
-extern tNFC_STATUS RW_I93UpdateNDef(uint16_t length, uint8_t* p_data);
+extern tNFC_STATUS RW_I93UpdateNDef(uint32_t length, uint8_t* p_data);
 
 /*******************************************************************************
 **
